renamed EmailItem.jsx, EmailList.jsx, SearchBar.jsx
