+++ src/components/EmailItem/index.jsx
